Give a sole module for buffer protocol
