+++ vm/src/stdlib/sre.rs
@@ -2,13 +2,8 @@ pub(crate) use _sre::make_module;
 
 #[pymodule]
 mod _sre {
-    use crossbeam_utils::atomic::AtomicCell;
-    use itertools::Itertools;
-    use num_traits::ToPrimitive;
-    use rustpython_common::hash::PyHash;
-
+    use crate::buffer::PyBufferRef;
     use crate::builtins::list::PyListRef;
-    use crate::builtins::memory::PyBufferRef;
     use crate::builtins::tuple::PyTupleRef;
     use crate::builtins::{
         PyCallableIterator, PyDictRef, PyInt, PyList, PyStr, PyStrRef, PyTypeRef,
@@ -21,6 +16,10 @@ mod _sre {
         PyValue, StaticType, TryFromBorrowedObject, TryFromObject,
     };
     use core::str;
+    use crossbeam_utils::atomic::AtomicCell;
+    use itertools::Itertools;
+    use num_traits::ToPrimitive;
+    use rustpython_common::hash::PyHash;
     use sre_engine::constants::SreFlag;
     use sre_engine::engine::{lower_ascii, lower_unicode, upper_unicode, State, StrDrive};
